+++ src/Contracts/LocalizedRoute.php
@@ -10,9 +10,6 @@ interface LocalizedRoute
     /** @see \Goodcat\L10n\Mixin\LocalizedRoute::lang */
     public function lang(): Route|RouteTranslations;
 
-    /** @see \Goodcat\L10n\Mixin\LocalizedRoute::getLocalizedName */
-    public function getLocalizedName(string $locale): ?string;
-
     /**
      * @return Route[]
      * @see \Goodcat\L10n\Mixin\LocalizedRoute::makeTranslations