+++ src/Routing/LocalizedUrlGenerator.php
@@ -2,44 +2,19 @@
 
 namespace Goodcat\L10n\Routing;
 
-use BackedEnum;
-use Goodcat\L10n\Contracts\LocalizedRoute;
-use Illuminate\Routing\Exceptions\UrlGenerationException;
 use Illuminate\Routing\UrlGenerator;
 use Illuminate\Support\Arr;
-use Illuminate\Support\Facades\Route;
-use InvalidArgumentException;
-use Symfony\Component\Routing\Exception\RouteNotFoundException;
 
 class LocalizedUrlGenerator extends UrlGenerator
 {
-    /**
-     * @throws UrlGenerationException
-     */
     public function route($name, $parameters = [], $absolute = true): string
     {
-        if ($name instanceof BackedEnum && ! is_string($name = $name->value)) {
-            throw new InvalidArgumentException('Attribute [name] expects a string backed enum.');
-        }
-
         $locale = Arr::pull($parameters, 'lang', app()->getLocale());
 
-        if (! is_null($route = $this->routes->getByName($name))) {
-            /** @var Route&LocalizedRoute  $route */
-            $localized = $route->getLocalizedName($locale);
-
-            if ($localized !== null && $localized !== $name) {
-                $route = $this->routes->getByName($localized);
-            }
-
-            return parent::toRoute($route, $parameters, $absolute);
+        if ($this->routes->hasNamedRoute("$name.$locale")) {
+            $name .= ".$locale";
         }
 
-        if (! is_null($this->missingNamedRouteResolver) &&
-            ! is_null($url = call_user_func($this->missingNamedRouteResolver, $name, $parameters, $absolute))) {
-            return $url;
-        }
-
-        throw new RouteNotFoundException("Route [{$name}] not defined.");
+        return parent::route($name, $parameters, $absolute);
     }
 }